对共因失效进行修正。

+++ ZGO2/include/itemgosignal.h
@@ -32,8 +32,8 @@ public:
     SignalConnection* end();
     void setEndPosition(QPointF pos);
     void setEndPosition(int x, int y);
-    double cornerX() const;
-    void setCornerX(const double value);
+    int cornerX() const;
+    void setCornerX(const int value);
     void paint(QPainter *painter, const QStyleOptionGraphicsItem *item, QWidget *widget);
     void updatePosition();
     void removeConnection();
@@ -45,7 +45,7 @@ protected:
     SignalConnection *_start;
     SignalConnection *_end;
     QPoint _endPos;
-    double _cornerX;
+    int _cornerX;
     GOSignal *_model;
 };
